Moved functions for assignment queue cards into hook to clean up assignmentqueuecards file
